Move broadcast to basae avtivity to not have to add it to every activity and just let base activity handle it.
