+++ app/src/main/java/com/brainbeats/MixerActivity.java
@@ -35,7 +35,6 @@ public class MixerActivity extends BaseActivity implements View.OnClickListener,
 
     Bundle mUserSelections;
     public FloatingActionButton mMainActionFab;
-    private IntentFilter mIntentFilter;
     public Mix mNewMix;
 
     @Override
@@ -52,9 +51,6 @@ public class MixerActivity extends BaseActivity implements View.OnClickListener,
             mMixerDetailFragment = new MixerDetailFragment();
             switchToMixerFragment();
         }
-
-        mIntentFilter = new IntentFilter();
-        mIntentFilter.addAction(Constants.SONG_COMPLETE_BROADCAST_ACTION);
 
         if (mUserSelections == null) {
             mUserSelections = new Bundle();
@@ -148,35 +144,4 @@ public class MixerActivity extends BaseActivity implements View.OnClickListener,
         getMenuInflater().inflate(R.menu.menu_global, menu);
         return true;
     }
-
-    @Override
-    public void onPause() {
-        unregisterReceiver(mReceiver);
-        super.onPause();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        registerReceiver(mReceiver, mIntentFilter);
-    }
-
-    private BroadcastReceiver mReceiver = new BroadcastReceiver() {
-        @Override
-        public void onReceive(Context context, Intent intent) {
-/*            if(intent.getAction().equals(Constants.SONG_COMPLETE_BROADCAST_ACTION)) {
-                Track newTrack = (Track) intent.getExtras().getParcelable(Constants.KEY_EXTRA_SELECTED_TRACK);
-                mCurrentSongTitle.setText(newTrack.getTitle());
-                if (newTrack.getArtworkURL() == null)
-                    mAlbumThumbnail.setImageResource(R.drawable.placeholder);
-                else
-                    Picasso.with(MixerActivity.this).load(newTrack.getArtworkURL()).into(mAlbumThumbnail);
-
-                mCurrentSongArtistName.setText(newTrack.getUser().getUsername());
-
-                //Update the current playing song in base activity to the song from this broadcast
-                mCurrentSong = newTrack;
-            }*/
-        }
-    };
 }
