+++ app/src/main/java/com/brainbeats/SocialActivity.java
@@ -33,9 +33,6 @@ public class SocialActivity extends BaseActivity implements SocialFragment.OnFra
         mSocialFragment = new SocialFragment();
         mUserProfileFragment = new UserProfileFragment();
         switchToSocialFragment();
-
-        mIntentFilter = new IntentFilter();
-        mIntentFilter.addAction(Constants.SONG_COMPLETE_BROADCAST_ACTION);
     }
 
     @Override
@@ -62,35 +59,4 @@ public class SocialActivity extends BaseActivity implements SocialFragment.OnFra
         getMenuInflater().inflate(R.menu.menu_global, menu);
         return true;
     }
-
-    @Override
-    public void onPause() {
-        unregisterReceiver(mReceiver);
-        super.onPause();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        registerReceiver(mReceiver, mIntentFilter);
-    }
-
-    private BroadcastReceiver mReceiver = new BroadcastReceiver() {
-        @Override
-        public void onReceive(Context context, Intent intent) {
-            if(intent.getAction().equals(Constants.SONG_COMPLETE_BROADCAST_ACTION)) {
-                Track newTrack = (Track) intent.getExtras().getParcelable(Constants.KEY_EXTRA_SELECTED_TRACK);
-                mCurrentSongTitle.setText(newTrack.getTitle());
-                if (newTrack.getArtworkURL() == null)
-                    mAlbumThumbnail.setImageResource(R.drawable.placeholder);
-                else
-                    Picasso.with(SocialActivity.this).load(newTrack.getArtworkURL()).into(mAlbumThumbnail);
-
-                mCurrentSongArtistName.setText(newTrack.getUser().getUsername());
-
-                //Update the current playing song in base activity to the song from this broadcast
-                mCurrentSong = newTrack;
-            }
-        }
-    };
 }
